Test

See merge request distribution/community-server-parent!52

+++ base-config/base-config-service/base-config-flow/src/main/java/cn/cuiot/dmp/baseconfig/flow/dto/approval/QueryMyApprovalDto.java
@@ -17,7 +17,7 @@ public class QueryMyApprovalDto extends PageQuery {
     /**
      * 业务类型
      */
-    private Byte businessType;
+    private Long businessType;
 
     /**
      * 审批人id
